Refactors "RichTextFieldView" to be child of abstract "ProseMirrorFieldView"
